Creación clase prueba

+++ src/Oferta.java
@@ -3,16 +3,16 @@ public final class Oferta {
     private Local local;
     private Articulo articulo;
     private String descripcion;
-    private int valor;
+    private double valor;
     private int vigencia;
     
-    public Oferta(Local elLocal,Articulo elArticulo,String laDescripcion,int elValor ,int laVigencia){
+    public Oferta(Local elLocal,Articulo elArticulo,String laDescripcion,double elValor ,int laVigencia){
                   
-                  this.setLocal(elLocal);
-                  this.setArticulo(elArticulo);
-                  this.setDescripcion(laDescripcion);
-                  this.setValor(elValor);
-                  this.setVigencia(laVigencia);
+        this.setLocal(elLocal);
+        this.setArticulo(elArticulo);
+        this.setDescripcion(laDescripcion);
+        this.setValor(elValor);
+        this.setVigencia(laVigencia);
    
     }
     
@@ -26,7 +26,7 @@ public final class Oferta {
     public String getDescripcion(){
         return descripcion;
     }
-    public int getValor(){
+    public double getValor(){
         return valor;
     }
     public int getVigencia(){
@@ -41,7 +41,7 @@ public final class Oferta {
     public void setDescripcion(String laDescripcion){
         descripcion=laDescripcion;
     }
-    public void setValor(int elValor){
+    public void setValor(double elValor){
         valor=elValor;
     }
     public void setVigencia(int laVigencia){
